FIX move test to test_parallel file+fix linter
